Swithc to material 3

+++ app/src/main/java/project/note/presentation/screens/NoteScreen.kt
@@ -1,48 +1,56 @@
 package project.note.presentation.screens
 
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.ClickableText
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.Divider
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.ModalBottomSheetLayout
-import androidx.compose.material.ModalBottomSheetValue
-import androidx.compose.material.Scaffold
-import androidx.compose.material.TextField
-import androidx.compose.material.TopAppBar
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.MoreVert
-import androidx.compose.material.rememberModalBottomSheetState
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import kotlinx.coroutines.launch
 import project.note.R
 import project.note.presentation.models.NoteViewModel
 
-@OptIn(ExperimentalMaterialApi::class)
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun NoteScreen(onBackClick: () -> Unit, viewModel: NoteViewModel = hiltViewModel()) {
-    val bottomSheetState =
-        rememberModalBottomSheetState(initialValue = ModalBottomSheetValue.Hidden)
     val coroutineScope = rememberCoroutineScope()
+    var showBottomSheet by rememberSaveable {
+        mutableStateOf(false)
+    }
 
     Scaffold(
         topBar = {
@@ -82,7 +90,7 @@ fun NoteScreen(onBackClick: () -> Unit, viewModel: NoteViewModel = hiltViewModel
 
                     IconButton(onClick = {
                         coroutineScope.launch {
-                            bottomSheetState.show()
+                            showBottomSheet = true
                         }
 
                     }) {
@@ -120,29 +128,55 @@ fun NoteScreen(onBackClick: () -> Unit, viewModel: NoteViewModel = hiltViewModel
                 })
             }
 
-            ModalBottomSheetLayout(
-                sheetState = bottomSheetState,
-                sheetContent = {
-                    LazyColumn(modifier = Modifier.padding(16.dp)) {
-                        items(1, itemContent = {
-                            ClickableText(modifier = Modifier
-                                .fillMaxWidth()
-                                .height(48.dp),
-                                text = AnnotatedString("Delete"),
-                                onClick = {
-                                    viewModel.delete()
-                                    onBackClick()
-                                })
-                            Divider(color = Color.Black, thickness = 1.dp)
-                        })
-                    }
-                },
-                content = {
-                    // Main content of the screen goes here
-                },
-                scrimColor = Color.Black.copy(alpha = 0.5f),
-                sheetShape = RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp)
-            )
+            if (showBottomSheet) {
+                CustomBottomSheet(onDeleteAction = {
+                    onBackClick()
+                    viewModel.delete()
+                }, onDismiss = {
+                    showBottomSheet = false
+                })
+            }
         }
     )
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun CustomBottomSheet(onDeleteAction: () -> Unit, onDismiss: () -> Unit) {
+    ModalBottomSheet(
+        modifier = Modifier,
+        sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = false),
+        onDismissRequest = {
+            onDismiss()
+        },
+        shape = RoundedCornerShape(
+            topStart = 10.dp,
+            topEnd = 10.dp
+        ),
+    ) {
+        CustomBottomSheetContainer(onDeleteAction)
+    }
+}
+
+@Composable
+fun CustomBottomSheetContainer(onDeleteAction: () -> Unit) {
+    Scaffold(topBar = {
+        Column {
+            Text(
+                text = "Note actions", modifier = Modifier
+                    .height(75.dp)
+                    .padding(start = 29.dp, top = 26.dp), fontSize = 23.sp
+            )
+            HorizontalDivider(thickness = 1.dp, color = Color.Black)
+        }
+    }) {
+        Column(modifier = Modifier.padding(it)) {
+            Text(text = "Delete",  textAlign = TextAlign.Center,  modifier = Modifier
+                .fillMaxWidth()
+                .height(48.dp)
+                .clickable {
+                    onDeleteAction()
+                })
+        }
+    }
 }
